w3c-tc-sample/test3.htm

<!DOCTYPE html>
<html>
  <head>
    <title>window.applicationCache</title>
    <link rel="help" href="http://www.w3.org/TR/2012/WD-css3-animations-20120403/#animation-direction-property" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if Animations.animation-direction has value reverse" />

    <script src="/resources/testharness.js" type="text/javascript"></script>
    <script src="/resources/testharnessreport.js" type="text/javascript"></script>
 
  </head>
  <body>
    <div id="log"></div>
    <div id="test"></div>
 
    <script type="text/javascript">
        navigator.getUserMedia = navigator.getUserMedia || navigator.webkitGetUserMedia;
        function gotUserMediaSuccess(s) {
            var url =!!webkitURL ? webkitURL.createObjectURL(s):window.URL.createObjectURL(s);
            document.getElementById("localView").src = url;
            assert_true(true, "getUserMedia SuccessCallback tests");
        }
        function gotUserMediaError(error) {
            assert_true(false, "It should enter gotUserMediaSuccess.");
        }
        function testUserMedia() {
            try {
                if (navigator.getUserMedia) {
                    navigator.getUserMedia({audio: true}, gotUserMediaSuccess, gotUserMediaError);
                } else {
                    assert_true(false, "No microphone installed in your computer");
                }
            } catch (e) {
                assert_true(false, "No microphone installed in your computer");
            }
        }
        test(testUserMedia, "Test Description: Test GetUserMedia SuccessCallback.");
    </script>
    </script>
  </body>
</html>
